add for inference function

+++ smartQ/routers/smartQ.py
@@ -1,11 +1,10 @@
-from typing import List
 from fastapi import APIRouter, Depends, status, HTTPException, Request
 
-from smartQ import schemas, oauth2
+from smartQ import schemas, oauth2, utils, rabbitmq, database
 
 router = APIRouter(
     prefix="/smartQ",
-    tags=['Blogs']
+    tags=['SmartQ']
 )
 
 
@@ -18,7 +17,7 @@ def all(db: Request, current_user: schemas.User = Depends(oauth2.get_current_use
     return list_blogs
 
 
-@router.post('/', status_code=status.HTTP_201_CREATED,)
+@router.post('/', status_code=status.HTTP_201_CREATED)
 def create(db: Request, request: schemas.Blog, current_user: schemas.User = Depends(oauth2.get_current_user)):
     new_blog = schemas.Blog(title=request.title, body=request.body) 
     dict_new_blog = dict(new_blog)
@@ -31,9 +30,30 @@ def create(db: Request, request: schemas.Blog, current_user: schemas.User = Depe
 # 아직 못함
 @router.delete('/{id}', status_code=status.HTTP_204_NO_CONTENT)
 def destroy(db: Request, id: int, current_user: schemas.User = Depends(oauth2.get_current_user)):
-    blog = db.app.database["Blogs"].delete_one({'id': id})
+    blog = db.app.database["Blogs"].delete_one({'email': current_user.email}, {'id': id})
     if not blog.first():
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Blog with id {id} not found")
         
     return 'done'
 
+"""
+추론
+request : current_user, img_file, list(model name), list(device_name)
+response : None, just do inference
+"""
+
+@router.get('/inference', status_code=status.HTTP_200_OK)
+def inference(inf_data: schemas.InferenceData, current_user: schemas.User = Depends(oauth2.get_current_user)):
+    
+    image = utils.extract_img(inf_data.image)
+    rabbitmq.publish(header='image', message=image, exchange_name='?', routing_key_name='?')
+
+    models = database.get_models(email=current_user.email, model_names=inf_data.model_names)
+    for model in models:
+        # model = {'model_name': resnet, 'model_contents': 1e23nfjui}
+        rabbitmq.publish(header='model', message=model, exchange_name='?', routing_key_name='?')
+        
+        
+        
+        
+    return True
